Lizard Tail Relayering
Previously, when facing upwards (away from the player), lizard tails would not be in the foreground in front of the clothes lizards were wearing. This adds a tail overlay to just that direction to genetics.dmi, which is called and applied in mutantraces.dm and update_icon.dm. Has been tested and it works.